Compiler options required for building with MSVC on Win x64
- The official define for Windows-detection is "_WIN32" not "WIN32".
This option is passed because 3rdparty/exprtk/exprtk.hpp is doing
windows detection for the incorrect "WIN32" define.

- Note that the Mingw compiler does define WIN32 in addition to
"_WIN32", but unfortunately, my gcc 4.7.0 fails with the "too many
sections" error in the object file.  The PE/COFF binary format has a
limit of 2^15 sections, which is too small for large template
libraries like exprTK.

- MSVC provides the "bigobj" switch for such cases, in which the
limit of sections is raised to 2^32.  This switch has also been
added, via extra_compile_args.